--- example/.editorconfig
@@ -1,3 +1,0 @@
-# Windows files
-[*.bat]
-end_of_line = crlf
