+++ example/android/app/src/debug/AndroidManifest.xml
@@ -1,16 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" >
+    xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.SYSTEM_ALERT_WINDOW"/>
-<uses-permission android:name="android.permission.INTERNET" />
     <application
         android:usesCleartextTraffic="true"
         tools:targetApi="28"
-        tools:ignore="GoogleAppIndexingWarning"
-        tools:replace="allowBackup"
-              android:allowBackup="false"
-        >
-        <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
-    </application>
+        tools:ignore="GoogleAppIndexingWarning"/>
 </manifest>
