+++ example/android/build.gradle
@@ -1,55 +1,21 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
     ext {
-        buildToolsVersion = "31.0.3"
-        minSdkVersion = 21
-        compileSdkVersion = 31
-        targetSdkVersion = 31
-        ndkVersion = "20.1.5948944"
+        buildToolsVersion = "34.0.0"
+        minSdkVersion = 23
+        compileSdkVersion = 34
+        targetSdkVersion = 34
+        ndkVersion = "26.1.10909125"
+        kotlinVersion = "1.9.22"
     }
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath("com.android.tools.build:gradle:4.1.0")
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
+        classpath("com.android.tools.build:gradle")
+        classpath("com.facebook.react:react-native-gradle-plugin")
+        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin")
     }
 }
 
-allprojects {
-    repositories {
-//        exclusiveContent {
-//            // We get React Native's Android binaries exclusively through npm,
-//            // from a local Maven repo inside node_modules/react-native/.
-//            // (The use of exclusiveContent prevents looking elsewhere like Maven Central
-//            // and potentially getting a wrong version.)
-//            filter {
-//                includeGroup "com.facebook.react"
-//            }
-//            forRepository {
-//                maven {
-//                    // NOTE: if you are in a monorepo, you may have "$rootDir/../../../node_modules/react-native/android"
-//                    url "$rootDir/../node_modules/react-native/android"
-//                }
-//            }
-//        }
-        mavenLocal()
-        maven {
-            // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
-            url("$rootDir/../node_modules/react-native/android")
-        }
-        maven {
-            // Android JSC is installed from npm
-            url("$rootDir/../node_modules/jsc-android/dist")
-        }
-
-        google()
-        jcenter()
-        maven { url 'https://www.jitpack.io' }
-    }
-}
-
-
+apply plugin: "com.facebook.react.rootproject"
