+++ example/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@react-native/typescript-config/tsconfig.json"
+}
